requirements.txt
langchain
langchain-community
langchain-huggingface
pypdf
faiss-cpu
sentence-transformers
streamlit
python-dotenv